requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain
langchain_community
flask
pypdf
python-dotenv
chainlit
streamlit

-e . # look setup.py   ->   # after running it egg.info folder generated